Service/requirements.txt
mysql-connector-python
flask
flask_cors
google-api-python-client
google-auth-oauthlib
msal
requests
